requirements.txt
Pillow==8.4.0
psycopg2==2.9.1
